+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.0
+* (Breaking changes) Rename `getPromise` to `service`
+* Add `resolve` method into mixin
+
 ## 1.0.1
 * Fix problem with renderToString when ResolveMixin is not used